Implement a Storable CSize instance
generatePrimes calls peek on a Ptr CSize, which requires a Storable
CSize instance. The current Storable CSize instance is empty, so
generatePrimes throws a runtime exception. We implement a proper
Storable CSize instance, via the Storable instance of the underlying
Word32/Word64.

Note: Implementing a Storable CSize requires a PrimType CSize instance,
and our PrimType CSize instance implementation is missing some methods
(which we don't use). This will eventually be a moot point, because
we'll switch to the identically-named Storable typeclass in base,
from the current Storable typeclass defined in basement. The Storable
from base has a proper Storable CSize instance.